AI/requirements.txt
# absl-py==2.1.0
# annotated-types==0.7.0
# anyio==4.3.0
# argon2-cffi==21.3.0
# argon2-cffi-bindings==21.2.0
# arrow==1.3.0
# asttokens==2.4.1
# astunparse==1.6.3
# async-lru==2.0.4
# attrs==23.2.0
# Babel==2.14.0
# beautifulsoup4==4.12.3
# bleach==6.1.0
# blinker==1.8.2
# certifi==2024.2.2
# cffi==1.16.0
# charset-normalizer==3.3.2
# click==8.1.7
# colorama==0.4.6
# comm==0.2.2
# contourpy==1.2.1
# cycler==0.12.1
# deap==1.4.1
# debugpy==1.8.1
# decorator==5.1.1
# defusedxml==0.7.1
# executing==2.0.1
# fastapi==0.95.0
# fastjsonschema==2.19.1
# Flask==3.0.3
# flatbuffers==24.3.25
# fonttools==4.51.0
# fqdn==1.5.1
# gast==0.5.4
# google-pasta==0.2.0
# grpcio==1.63.0
# gunicorn==23.0.0
# h11==0.14.0
# h5py==3.11.0
# httpcore==1.0.5
# httpx==0.27.0
# idna==3.7
# imbalanced-learn==0.12.3
# ipykernel==6.29.4
# ipywidgets==8.1.2
# isoduration==20.11.0
# itsdangerous==2.2.0
# jedi==0.19.1
# Jinja2==3.1.3
# joblib==1.4.0
# json5==0.9.25
# jsonpointer==2.4
# jsonschema==4.22.0
# jsonschema-specifications==2023.12.1
# jupyter==1.0.0
# jupyter-console==6.6.3
# jupyter-events==0.10.0
# jupyter-lsp==2.2.5
# jupyter_client==8.6.1
# jupyter_core==5.7.2
# jupyter_server==2.14.0
# jupyter_server_terminals==0.5.3
# jupyterlab==4.1.8
# jupyterlab_pygments==0.3.0
# jupyterlab_server==2.27.1
# jupyterlab_widgets==3.0.10
# keras==3.3.3
# kiwisolver==1.4.5
# libclang==18.1.1
# lightgbm==4.5.0
# lxml==5.3.0
# Markdown==3.6
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# matplotlib==3.8.4
# matplotlib-inline==0.1.7
# mdurl==0.1.2
# mistune==3.0.2
# ml-dtypes==0.3.2
# namex==0.0.8
# nbclient==0.10.0
# nbconvert==7.16.4
# nbformat==5.10.4
# nest-asyncio==1.6.0
# nltk==3.8.1
# notebook==7.1.3
# notebook_shim==0.2.4
# numpy==1.26.4
# opt-einsum==3.3.0
# optree==0.11.0
# overrides==7.7.0
# packaging==24.0
# pandas==2.2.2
# pandocfilters==1.5.1
# parso==0.8.4
# pillow==10.3.0
# platformdirs==4.2.1
# prometheus_client==0.20.0
# prompt-toolkit==3.0.43
# protobuf==4.25.3
# psutil==5.9.8
# pure-eval==0.2.2
# pycparser==2.22
# pydantic==1.10.7
# pydantic_core==2.27.2
# Pygments==2.17.2
# pyparsing==3.1.2
# python-dateutil==2.9.0.post0
# python-docx==1.1.2
# python-json-logger==2.0.7
# python-multipart==0.0.20
# python-pptx==1.0.2
# pytube==15.0.0
# pytz==2024.1
# #pywin32==306
# pywinpty==2.0.13
# PyYAML==6.0.1
# pyzmq==26.0.3
# qtconsole==5.5.1
# QtPy==2.4.1
# referencing==0.35.1
# regex==2024.4.28
# requests==2.31.0
# rfc3339-validator==0.1.4
# rfc3986-validator==0.1.1
# rich==13.7.1
# rpds-py==0.18.0
# scikit-learn==1.4.2
# scipy==1.13.0
# seaborn==0.13.2
# Send2Trash==1.8.3
# setuptools==69.5.1
# six==1.16.0
# sniffio==1.3.1
# soupsieve==2.5
# stack-data==0.6.3
# starlette==0.45.3
# tensorboard==2.16.2
# tensorboard-data-server==0.7.2
# tensorflow-intel==2.16.1
# termcolor==2.4.0
# terminado==0.18.1
# tflearn==0.5.0
# threadpoolctl==3.5.0
# tinycss2==1.3.0
# tornado==6.4
# tqdm==4.66.2
# traitlets==5.14.3
# types-python-dateutil==2.9.0.20240316
# typing_extensions==4.12.2
# tzdata==2024.1
# uri-template==1.3.0
# urllib3==2.2.1
# uvicorn==0.21.1
# wcwidth==0.2.13
# webcolors==1.13
# webencodings==0.5.1
# websocket-client==1.8.0
# Werkzeug==3.0.2
# wheel==0.43.0
# widgetsnbextension==4.0.10
# wrapt==1.16.0
# xgboost==2.1.1
# XlsxWriter==3.2.0
# yt-dlp==2024.12.13

fastapi==0.95.0
uvicorn==0.21.1
tensorflow>=2.10,<3.0
numpy==1.26.4
Pillow==10.3.0
python-multipart==0.0.20
requests==2.31.0
httpx==0.27.0